no-framework/crm-vendas/crm/requirements.txt
streamlit 
pydantic[email] 
psycopg2-binary 
python-dotenv
mkdocs 
mkdocs-material 
mkdocstrings
mkdocstrings-python